sync 20211217

+++ procon-services/system-service/src/main/java/com/huangpuguang/system/mapper/SysMenuMapper.java
@@ -55,7 +55,7 @@ public interface SysMenuMapper
     /**
      * 根据用户ID查询菜单
      *
-     * @param username 用户ID
+     * @param userId 用户ID
      * @return 菜单列表
      */
     public List<SysMenu> selectMenuTreeByUserId(Long userId);
@@ -67,7 +67,7 @@ public interface SysMenuMapper
      * @param menuCheckStrictly 菜单树选择项是否关联显示
      * @return 选中菜单列表
      */
-    public List<Integer> selectMenuListByRoleId(@Param("roleId") Long roleId, @Param("menuCheckStrictly") boolean menuCheckStrictly);
+    public List<Long> selectMenuListByRoleId(@Param("roleId") Long roleId, @Param("menuCheckStrictly") boolean menuCheckStrictly);
 
     /**
      * 根据菜单ID查询信息
