+++ procon-services/system-service/src/main/java/com/huangpuguang/system/service/impl/SysDeptServiceImpl.java
@@ -101,7 +101,7 @@ public class SysDeptServiceImpl implements SysDeptService
      * @return 选中部门列表
      */
     @Override
-    public List<Integer> selectDeptListByRoleId(Long roleId)
+    public List<Long> selectDeptListByRoleId(Long roleId)
     {
         SysRole role = roleMapper.selectRoleById(roleId);
         return deptMapper.selectDeptListByRoleId(roleId, role.isDeptCheckStrictly());
